+++ packages/orq-rc/RELEASES.md
@@ -9159,3 +9159,13 @@ Based on:
 - [typescript v4.10.0-rc.1] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.1 - packages/orq-rc
+
+## 2026-05-13 12:12:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.2] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.2 - packages/orq-rc